api/pyproject.toml
[tool.poetry]
name = "docker-template"
version = "1.0.0"
description = "Docker Template with FastAPI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asyncio = "^3.4.3"
debugpy = "~=1.8.13"
fastapi = "~=0.115.12"
httpx = "~=0.28.1"
python-dotenv = "~=1.1.0"
uvicorn = {extras = ["standard"], version = "~=0.34.0"}
websockets = "~=15.0.1"
openai = "^1.77.0"
langchain-openai = "^0.3.16"
langchain-core = "^0.3.59"
python-slugify = "^8.0.4"
numpy = "^2.2.5"
zep-cloud = "^2.12.1"
colorama = "^0.4.6"
langchain = "^0.3.25"
langgraph = "^0.4.3"
freeplay = "^0.3.21"
gradio = "^5.29.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"